Fixes part of issue #1951 pertaining to schedules.
This code was not using the universal set() method, instead setting the properties
of Occurence directly.  This caused ORDataObject::_getTimestamp() to have a string
instead of a TimestampObject which created the non-object error.

Also addressed in this is an unknown index error if the start_date and/or end_date
were not set.  If they weren't set, an error message would be registered, but it
would still attempt to create a time from the string resulting in the error.