This gets an encount total onto the report under 'Total Encounters', but doesn't address the underlying cause.
The original query start an Occurence and connects up to the Encounter, while this starts at the Encounter and moves to the Occurence, if it exists.  I'm not sure how the Encouter/Occurence coupling is supposed to work, so I'll have to wait until I can pick someone else's brain on this before proceeding any further.