A refractoring to minimize the coupling with $this->get().
The ID is accessible via the ORDO once the ORDO is loaded, so I created the _loadPatient() to handle pulling/putting the ID into the Session as well as generate the Patient ORDO and returning it.  After that, all mentions of $patient_id and/or $this->get('patient_id'...) have been changed to $p->get('id').