Adds support for primary diagnosis per final part of 4736.
This doesn't seem to work as you'd expect.  Somehow the left join on coding_data is adding more results than without.  There's a double join in there somewhere, or possibly some bad data in one table or another.  The group by at the end fixes it by returning the same number of results as without the the join.  There appears to be bad data in coding_data.