Contains a bunch of changes all aimed at issue #4475
Updates PatientStatistics to utilize new value() method for displaying formatted data instead of raw data
Adds in PatientStatistics controller and actionView() to handle display
Adds actionView() results to dashboard display.

This is a semi-experimental commit to see what the general concensus is about this style of design.  With this, all of our 'collection' controllers like PatientDashboard, would call ordo-specific controllers for each display block.  This makes it easy to pull any one block out of one page and re-use it elsewhere, but it does add a little dectective work to find out where a block is being generated and might make the structure a little more complex for new programmers to wrap their heads around.  Comments and such are welcomed.